Create initial classes and concepts for library
Add classes for semantic versions, public apis, calls, and call chains.
Does a lot of walking and filtering of modules, classes, and functions.
Currently supports comparing two modules in different folders.  This
can be achieved by installing or copying the code into that other
folder and then providing the path to the module.

Still to do:

 - Add tests and comments and documentation.

 - Add other ways to examine APIs, maybe through whl files.

 - Ensure it's usable in its own setup.py file.